requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asgi-redis==1.4.2
asgiref==1.1.2            # via asgi-redis, channels, daphne
asn1crypto==0.22.0        # via cryptography
attrs==17.2.0             # via automat, service-identity
autobahn==17.7.1          # via daphne
automat==0.6.0            # via twisted
beautifulsoup4==4.6.0
certifi==2017.7.27.1      # via requests
cffi==1.10.0              # via cryptography
channels==1.1.6
chardet==3.0.4            # via requests
constantly==15.1.0        # via twisted
cryptography==2.0.3       # via pyopenssl
daphne==1.3.0             # via channels
django-autoslug==1.9.3
django-bootstrap3==9.0.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django==1.11.4
djangorestframework==3.6.3
h2==3.0.1                 # via twisted
hpack==3.0.0              # via h2
html2text==2016.9.19
hyperframe==5.1.0         # via h2
hyperlink==17.3.0         # via twisted
idna==2.5                 # via cryptography, requests, twisted
incremental==17.5.0       # via twisted
msgpack-python==0.4.8     # via asgi-redis
olefile==0.44             # via pillow
pillow==4.2.1
priority==1.3.0           # via twisted
pyasn1-modules==0.0.11    # via service-identity
pyasn1==0.3.2             # via pyasn1-modules, service-identity
pycparser==2.18           # via cffi
pyopenssl==17.2.0         # via service-identity, twisted
pytz==2017.2              # via django
redis==2.10.5             # via asgi-redis
requests==2.18.3
service-identity==17.0.0  # via twisted
six==1.10.0               # via asgi-redis, asgiref, autobahn, automat, cryptography, pyopenssl, txaio
twisted[http2,tls]==17.5.0
txaio==2.8.1              # via autobahn
urllib3==1.22             # via requests
zope.interface==4.4.2     # via twisted
